+++ modules/main/src/main/java/com/nisovin/shopkeepers/commands/shopkeepers/snapshot/CommandSnapshotList.java
@@ -7,6 +7,8 @@ import org.bukkit.command.CommandSender;
 
 import com.nisovin.shopkeepers.api.ShopkeepersPlugin;
 import com.nisovin.shopkeepers.api.shopkeeper.ShopkeeperSnapshot;
+import com.nisovin.shopkeepers.api.shopkeeper.player.PlayerShopkeeper;
+import com.nisovin.shopkeepers.api.shopkeeper.player.members.DefaultPlayerShopAccessLevels;
 import com.nisovin.shopkeepers.api.ui.DefaultUITypes;
 import com.nisovin.shopkeepers.commands.arguments.ShopkeeperArgument;
 import com.nisovin.shopkeepers.commands.arguments.ShopkeeperFilter;
@@ -20,6 +22,7 @@ import com.nisovin.shopkeepers.commands.util.ShopkeeperArgumentUtils.TargetShopk
 import com.nisovin.shopkeepers.config.Settings.DerivedSettings;
 import com.nisovin.shopkeepers.lang.Messages;
 import com.nisovin.shopkeepers.shopkeeper.AbstractShopkeeper;
+import com.nisovin.shopkeepers.shopkeeper.player.AbstractPlayerShopkeeper;
 import com.nisovin.shopkeepers.util.bukkit.TextUtils;
 
 class CommandSnapshotList extends Command {
@@ -53,7 +56,12 @@ class CommandSnapshotList extends Command {
 		AbstractShopkeeper shopkeeper = context.get(ARGUMENT_SHOPKEEPER);
 		int page = context.get(ARGUMENT_PAGE);
 
-		if (!shopkeeper.canEdit(sender, false)) {
+		// Check access:
+		if (shopkeeper instanceof PlayerShopkeeper) {
+			if (!((AbstractPlayerShopkeeper) shopkeeper).checkAccess(sender, DefaultPlayerShopAccessLevels.FULL(), false)) {
+				return;
+			}
+		} else if (!shopkeeper.canEdit(sender, false)) {
 			return;
 		}