+++ modules/main/src/main/java/com/nisovin/shopkeepers/commands/shopkeepers/snapshot/CommandSnapshotRestore.java
@@ -8,6 +8,8 @@ import com.nisovin.shopkeepers.api.ShopkeepersPlugin;
 import com.nisovin.shopkeepers.api.shopkeeper.Shopkeeper;
 import com.nisovin.shopkeepers.api.shopkeeper.ShopkeeperLoadException;
 import com.nisovin.shopkeepers.api.shopkeeper.ShopkeeperSnapshot;
+import com.nisovin.shopkeepers.api.shopkeeper.player.PlayerShopkeeper;
+import com.nisovin.shopkeepers.api.shopkeeper.player.members.DefaultPlayerShopAccessLevels;
 import com.nisovin.shopkeepers.api.ui.DefaultUITypes;
 import com.nisovin.shopkeepers.commands.arguments.ShopkeeperArgument;
 import com.nisovin.shopkeepers.commands.arguments.ShopkeeperFilter;
@@ -22,6 +24,7 @@ import com.nisovin.shopkeepers.commands.util.ShopkeeperArgumentUtils.TargetShopk
 import com.nisovin.shopkeepers.config.Settings.DerivedSettings;
 import com.nisovin.shopkeepers.lang.Messages;
 import com.nisovin.shopkeepers.shopkeeper.AbstractShopkeeper;
+import com.nisovin.shopkeepers.shopkeeper.player.AbstractPlayerShopkeeper;
 import com.nisovin.shopkeepers.util.bukkit.TextUtils;
 import com.nisovin.shopkeepers.util.logging.Log;
 
@@ -57,7 +60,12 @@ class CommandSnapshotRestore extends Command {
 		assert snapshotIndex >= 0 && snapshotIndex < shopkeeper.getSnapshots().size();
 		int snapshotId = snapshotIndex + 1;
 
-		if (!shopkeeper.canEdit(sender, false)) {
+		// Check access:
+		if (shopkeeper instanceof PlayerShopkeeper) {
+			if (!((AbstractPlayerShopkeeper) shopkeeper).checkAccess(sender, DefaultPlayerShopAccessLevels.FULL(), false)) {
+				return;
+			}
+		} else if (!shopkeeper.canEdit(sender, false)) {
 			return;
 		}