+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopkeeper/AbstractShopkeeper.java
@@ -1755,7 +1755,7 @@ public abstract class AbstractShopkeeper implements Shopkeeper {
 	 *            the {@link UIType}, not <code>null</code>
 	 * @param silent
 	 *            <code>true</code> to omit any feedback that might otherwise be sent to the sender
-	 * @return <code>true</code> if the sender can access the UI this shopkeeper
+	 * @return <code>true</code> if the sender can access the specified UI type of this shopkeeper
 	 */
 	public final boolean canAccess(CommandSender sender, UIType uiType, boolean silent) {
 		var viewProvider = this.getViewProvider(uiType);